+++ R/XGR_plot_style.R
@@ -1,10 +1,10 @@
 XGR_plot_style <- function(xgr_track,
                            gr.filt,
                            lib_name,
-                           palette = c("Spectral", "BrBG", "PiYG", "PuOr")) {
-    requireNamespace("ggplot2")
-    #### further modify ####
-    colourCount <- length(unique(gr.filt$Assay))
+                           colors) {
+    
+    requireNamespace("ggplot2") 
+    #### further modify #### 
     trk_name <- paste("XGR:", lib_name)
     xgr_track <- suppressMessages(
         xgr_track +
@@ -13,11 +13,7 @@ XGR_plot_style <- function(xgr_track,
                 strip.text.y = ggplot2::element_text(angle = 0),
                 strip.text = ggplot2::element_text(size = 9)
             ) +
-            ggplot2::scale_fill_manual(
-                values = grDevices::colorRampPalette(
-                    RColorBrewer::brewer.pal(8, name = palette[1])
-                )(colourCount)
-            ) +
+            ggplot2::scale_fill_manual(values = colors) +
             ggplot2::scale_y_continuous(n.breaks = 3) +
             ggplot2::guides(fill = ggplot2::guide_legend(
                 ncol = 2,